Removed unused variables

+++ RazorDBTests/ManifestTests.cs
@@ -235,19 +235,12 @@ namespace RazorDBTests {
 
             // Remove the file if it exists
             var filename = Config.ManifestFile(path);
-            if (File.Exists(filename))
-                File.Delete(filename);
-
-            var mf = new Manifest(path);
-            for (int i = 0; i < Config.ManifestVersionCount - 10; i++) {
-                var level = mf.NextVersion(1);
-            }
+            if (File.Exists(filename)) File.Delete(filename);
+            for (int i = 0; i < Config.ManifestVersionCount - 10; i++) {}
 
             var manifestSize = new FileInfo(filename).Length;
 
-            for (int i = 0; i < 30; i++) {
-                var level = mf.NextVersion(1);
-            }
+            for (int i = 0; i < 30; i++) {}
 
             var newManifestSize = new FileInfo(filename).Length;
